Build in a pure Wayland environment
Note that this change only allow the package to build in a pure wayland
environment but doesn't make it functionnal since some X11 -> Wayland
equivalents doesn't exist yet. The changes that remain to be made in
order to get it working in a pure Wayland environment are marked by the
comment "TO DO WAYLAND".

Change-Id: Id97213ad446d6b8252bdf1447c8b8d4734443052
Signed-off-by: Kévin THIERRY <kevin.thierry@open.eurogiciel.org>